House the canvas stop tests with the keyboard navigation suite
Two tests do not carry a file; the suite about focus order is where
they belong.

Linters ran by hand; the pre-commit hook does not survive this environment.
